# replace string equality with charCode equality