Testing benchmark on node 7.10 : 

Before : 

lookup static route x 14,919,746 ops/sec ±0.77% (90 runs sampled)
lookup dynamic route x 662,481 ops/sec ±0.63% (94 runs sampled)
lookup long static route x 2,311,837 ops/sec ±0.98% (90 runs sampled)
find static route x 28,791,481 ops/sec ±0.88% (87 runs sampled)
find dynamic route x 732,577 ops/sec ±0.64% (94 runs sampled)
find long static route x 6,701,436 ops/sec ±0.75% (91 runs sampled)

After :

lookup static route x 16,330,523 ops/sec ±0.81% (91 runs sampled)
lookup dynamic route x 980,184 ops/sec ±0.61% (94 runs sampled)
lookup long static route x 3,957,358 ops/sec ±0.70% (90 runs sampled)
find static route x 29,006,918 ops/sec ±0.83% (94 runs sampled)
find dynamic route x 1,047,862 ops/sec ±0.53% (92 runs sampled)
find long static route x 6,818,069 ops/sec ±0.73% (91 runs sampled)