@billouboq 2017-05-08
@delvedor I see no improvement with this particular change.