@delvedor 2017-05-08
That's strange!
Since it helps here
```js
 while (i < req.url.length && req.url.charCodeAt(i) !== 63 /* ? */ && req.url.charCodeAt(i) !== 35 /* # */) i++
```
I was expecting the same result. Thanks for having tried it!